Fix inclusion order
Tailwind was getting included at the end of vendor.css, making it
impossible for addons to override or @extend Tailwind classes.

I am still not completely sure how the Tailwind file ends up in the
build, because it seems to work without any sort of @import statement
on the addon's behalf. Perhaps all *.css files under `/addon` are
included automatically.

In any case, tweaking the location of the built Tailwind output
file made it so it was included at the beginning, thus allowing
addons to override Tailwind classes.